refactor(controller): servlet 명명 대신 controller 사
