+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,7 +1,7 @@
 package org.apache.catalina;
 
 import java.io.IOException;
-import org.apache.catalina.servlets.http.Session;
+import org.apache.catalina.controller.http.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
