+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,0 +1,9 @@
+package org.apache.catalina.controller;
+
+import java.io.IOException;
+import org.apache.catalina.controller.http.request.HttpRequest;
+import org.apache.catalina.controller.http.response.HttpResponse;
+
+public interface Controller {
+    void service(HttpRequest request, HttpResponse response) throws IOException;
+}
