+++ tomcat/src/main/java/org/apache/catalina/controller/http/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.servlets.http;
+package org.apache.catalina.controller.http;
 
 public class Cookie {
