+++ tomcat/src/main/java/org/apache/catalina/controller/http/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.servlets.http;
+package org.apache.catalina.controller.http;
 
 import java.io.IOException;
 import java.util.HashMap;
