+++ tomcat/src/main/java/org/apache/catalina/controller/http/request/HttpMethod.java
@@ -1,0 +1,7 @@
+package org.apache.catalina.controller.http.request;
+
+public enum HttpMethod {
+
+    GET,
+    POST;
+}
