+++ tomcat/src/main/java/org/apache/catalina/controller/http/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.servlets.http.response;
+package org.apache.catalina.controller.http.response;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
