--- tomcat/src/main/java/org/apache/catalina/servlets/http/request/HttpMethod.java
@@ -1,7 +1,0 @@
-package org.apache.catalina.servlets.http.request;
-
-public enum HttpMethod {
-
-    GET,
-    POST;
-}
